feature gate behind serde-simple
