+++ README.md
@@ -19,6 +19,10 @@ cargo add ninterp
   ```
   cargo add ninterp --features serde
   ```
+- `serde-simple`: same as `serde` feature, with alternate simplified array serde format
+  ```
+  cargo add ninterp --features serde-simple
+  ```
 
 ## Examples
 See examples in `new` method documentation:
